+++ Samples~/Scripts/DemoExample/Core/ExampleManager.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Text.Json;
+using Beacon.Sdk.Beacon.Permission;
 using Beacon.Sdk.Beacon.Sign;
 using Netezos.Encoding;
 using TezosSDK.Beacon;
 using TezosSDK.Helpers;
-using TezosSDK.Scripts.FileUploaders;
-using TezosSDK.Scripts.IpfsUploader;
 using TezosSDK.Tezos;
 using TezosSDK.Tezos.API.Models;
 using TezosSDK.Tezos.API.Models.Tokens;
@@ -36,7 +35,18 @@ namespace TezosSDK.Samples.DemoExample
 
         public void Init(Action<bool> callback = null)
         {
-            Tezos = TezosSingleton.Instance;
+            Tezos = TezosSingleton.ConfiguredInstance(
+                networkType: NetworkType.ghostnet,
+                rpcUrl: "https://rpc.tzkt.io/ghostnet",
+                dAppMetadata: new DAppMetadata
+                {
+                    Name = "Tezos Demo Example Game",
+                    Url = "https://github.com/trilitech/tezos-unity-sdk/tree/main/Samples~",
+                    Icon =
+                        "https://raw.githubusercontent.com/" +
+                        "trilitech/tezos-unity-sdk/main/Samples%7E/Art/UI/Icons/skill_icon_04.png"
+                });
+
             _networkRPC = TezosConfig.Instance.RpcBaseUrl;
         }