+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [2.0.6]
+## [2.0.7] - 2023-10-24
+### Added
+- Configurable parameters (`networkType`, `rpcUrl`, `dAppMetadata` and `logLevel`) for Tezos singleton instance.  
+
+### Changed
+- Default RPC node to tezos.marigold.dev
+
+## [2.0.6] - 2023-10-16
 ### Changed
 - Disabled `Kukai Embed` silent signing
 - Updated `Kukai Embed` to `0.8.9`
@@ -162,7 +169,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added auto releases with GH actions
 
 
-[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/2.0.6...HEAD
+[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/2.0.7...HEAD
+[2.0.7]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/2.0.7
 [2.0.6]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/2.0.6
 [2.0.5]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/2.0.5
 [2.0.4]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/2.0.4